Move plugins to top level import and make setup optional
