+++ distributed/__init__.py
@@ -4,6 +4,7 @@ from .actor import Actor, ActorFuture
 from .core import connect, rpc
 from .deploy import LocalCluster, Adaptive, SpecCluster
 from .diagnostics.progressbar import progress
+from .diagnostics.plugin import WorkerPlugin, SchedulerPlugin
 from .client import (
     Client,
     Executor,
